[logic][l]: - create or update package moved into logic - tests moved into logic - mocks updated where appropriate - imports updated -refs #401
